Preparing the create thread page🖌️
Now I`m focused on the design and basic functionality of the thread create page which will have text editor and other useful things.Tomorrow she will be finished ,but there is a lot of work on her.
Other things is that the back-end has the needed Entities to start the new feature- thread ,but it will be uploaded at later time.