Create Thread Page & Back-end update
For now i think the create thread page from front-end is finished.
Started testing and implementing the functionality for creating a thread which will work with ajax ,because  my custom text editor can't send the data with submit button.
There is a lot of work with this piece of the project ,but now I don't have enought time to give everything from myself on this project :/