fixed the biggest bug to now
The bug was the following slow jms sendAndReceive which was caused by too big timeout 1s .I have changed it to 100ms which is enough to handle over 1500 request ,in future if they get too big it can be increased to 200ms which will double the requests amount.The next thing is the old messages in the queue which were sended without JMS Listener and they caused amazing slow and buggy trouble which was that the consumer cant return them ,which is normal ,so to fix this useless shi*t i have imeplement error handler for the jms of each component ,which will just remove them. For my application this fix is perfeect ,because I dont need to handle sendAndReceive data after some time.